+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Improvements:
 
 ## Bug fixes:
+* Enhance BoDi error handling to provide the name of the Interface being registered when that Interface has already been resolved.
 
 *Contributors of this release (in alphabetical order):*